+++ docs/src/examples/components/Input/Variations/InputExampleClearable.shorthand.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Input } from '@stardust-ui/react'
 
-const InputExampleClearableShorthand = () => <Input clearable placeholder="Search..." />
+const InputExampleClearable = () => <Input clearable placeholder="Search..." />
 
-export default InputExampleClearableShorthand
+export default InputExampleClearable