+++ src/themes/teams/components/Icon/svg/icons/index.ts
@@ -12,6 +12,11 @@ import more from './more'
 import teamCreate from './teamCreate'
 import umbrella from './umbrella'
 import teams from './teams'
+import format from './format'
+import giphy from './giphy'
+import fontColor from './fontColor'
+import fontSize from './fontSize'
+import highlight from './highlight'
 
 export default {
   call,
@@ -26,4 +31,9 @@ export default {
   'team-create': teamCreate,
   umbrella,
   teams,
+  format,
+  giphy,
+  highlight,
+  'font-color': fontColor,
+  'font-size': fontSize,
 } as { [iconName: string]: TeamsSvgIconSpec }