+++ test/specs/components/Chat/Chat-test.ts
@@ -1,10 +1,10 @@
 import { handlesAccessibility, isConformant } from 'test/specs/commonTests'
 
-import Chat from 'src/components/Chat'
+import Chat from 'src/components/Chat/Chat'
 import implementsCollectionShorthandProp from '../../commonTests/implementsCollectionShorthandProp'
 import ChatItem from 'src/components/Chat/ChatItem'
 import { chatBehavior } from 'src/lib/accessibility'
-import { IAccessibilityDefinition } from 'src/lib/accessibility/interfaces'
+import { AccessibilityDefinition } from 'src/lib/accessibility/types'
 
 const chatImplementsCollectionShorthandProp = implementsCollectionShorthandProp(Chat)
 
@@ -15,7 +15,7 @@ describe('Chat', () => {
   describe('accessibility', () => {
     handlesAccessibility(Chat, {
       defaultRootRole: 'presentation',
-      focusZoneDefinition: (chatBehavior as IAccessibilityDefinition).focusZone,
+      focusZoneDefinition: (chatBehavior as AccessibilityDefinition).focusZone,
     })
   })
 })