+++ test/specs/components/Menu/MenuItem-test.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react'
 
 import { isConformant, handlesAccessibility, getRenderedAttribute } from 'test/specs/commonTests'
-import { getTestingRenderedComponent } from 'test/utils'
+import { mountWithProviderAndGetComponent } from 'test/utils'
 import MenuItem from 'src/components/Menu/MenuItem'
 import { toolbarButtonBehavior, tabBehavior } from '../../../../src/lib/accessibility'
 
@@ -13,7 +13,7 @@ describe('MenuItem', () => {
   })
 
   it('content renders as `li > a`', () => {
-    const menuItem = getTestingRenderedComponent(MenuItem, <MenuItem content="Home" />).find(
+    const menuItem = mountWithProviderAndGetComponent(MenuItem, <MenuItem content="Home" />).find(
       '.ui-menu__item',
     )
 
@@ -23,7 +23,7 @@ describe('MenuItem', () => {
   })
 
   it('children render directly inside `li`', () => {
-    const menuItem = getTestingRenderedComponent(MenuItem, <MenuItem>Home</MenuItem>)
+    const menuItem = mountWithProviderAndGetComponent(MenuItem, <MenuItem>Home</MenuItem>)
 
     expect(menuItem.find('.ui-menu__item').is('li')).toBe(true)
     expect(menuItem.text()).toBe('Home')
@@ -35,18 +35,18 @@ describe('MenuItem', () => {
 
     describe('as a default MenuItem', () => {
       test('root role should be presentation', () => {
-        const menuItemComponent = getTestingRenderedComponent(MenuItem, <MenuItem />)
+        const menuItemComponent = mountWithProviderAndGetComponent(MenuItem, <MenuItem />)
         expect(getRenderedAttribute(menuItemComponent, 'role', '')).toBe('presentation')
       })
 
       test('anchor role should be menuitem', () => {
-        const menuItemComponent = getTestingRenderedComponent(MenuItem, <MenuItem />)
+        const menuItemComponent = mountWithProviderAndGetComponent(MenuItem, <MenuItem />)
         expect(getRenderedAttribute(menuItemComponent, 'role', 'a')).toBe('menuitem')
       })
 
       test('aria-label should be added to the child anchor', () => {
         const ariaLabel = 'Useful Tool Tip'
-        const menuItemComponent = getTestingRenderedComponent(
+        const menuItemComponent = mountWithProviderAndGetComponent(
           MenuItem,
           <MenuItem aria-label={ariaLabel} />,
         )
@@ -57,7 +57,7 @@ describe('MenuItem', () => {
 
       test('aria-labelledby should be added to the child anchor', () => {
         const ariaLabelledByID = 'element-that-labels'
-        const menuItemComponent = getTestingRenderedComponent(
+        const menuItemComponent = mountWithProviderAndGetComponent(
           MenuItem,
           <MenuItem aria-labelledby={ariaLabelledByID} accessibility={toolbarButtonBehavior} />,
         )
@@ -71,7 +71,7 @@ describe('MenuItem', () => {
 
     describe('as a ToolbarButton', () => {
       test('root role should be presentation', () => {
-        const menuItemComponent = getTestingRenderedComponent(
+        const menuItemComponent = mountWithProviderAndGetComponent(
           MenuItem,
           <MenuItem accessibility={toolbarButtonBehavior} />,
         )
@@ -79,7 +79,7 @@ describe('MenuItem', () => {
       })
 
       test('anchor role should be button', () => {
-        const menuItemComponent = getTestingRenderedComponent(
+        const menuItemComponent = mountWithProviderAndGetComponent(
           MenuItem,
           <MenuItem accessibility={toolbarButtonBehavior} />,
         )
@@ -88,7 +88,7 @@ describe('MenuItem', () => {
 
       test('aria-label should be added to the child anchor', () => {
         const ariaLabel = 'Useful Tool Tip'
-        const menuItemComponent = getTestingRenderedComponent(
+        const menuItemComponent = mountWithProviderAndGetComponent(
           MenuItem,
           <MenuItem aria-label={ariaLabel} accessibility={toolbarButtonBehavior} />,
         )
@@ -99,7 +99,7 @@ describe('MenuItem', () => {
 
       test('aria-labelledby should be added to the child anchor', () => {
         const ariaLabelledByID = 'element-that-labels'
-        const menuItemComponent = getTestingRenderedComponent(
+        const menuItemComponent = mountWithProviderAndGetComponent(
           MenuItem,
           <MenuItem aria-labelledby={ariaLabelledByID} accessibility={toolbarButtonBehavior} />,
         )
@@ -113,7 +113,7 @@ describe('MenuItem', () => {
       const disabledFlags = [true, false]
       disabledFlags.forEach(disabledValue => {
         test(`aria-disabled should be ${disabledValue} if menuitem disabled prop is ${disabledValue}`, () => {
-          const menuItemComponent = getTestingRenderedComponent(
+          const menuItemComponent = mountWithProviderAndGetComponent(
             MenuItem,
             <MenuItem disabled={disabledValue} accessibility={toolbarButtonBehavior} />,
           )
@@ -127,7 +127,7 @@ describe('MenuItem', () => {
 
       disabledFlags.forEach(disabledValue => {
         test(`aria-disabled should be ${disabledValue} if menuitem also has the disabled prop and aria-disabled is ${disabledValue}`, () => {
-          const menuItemComponent = getTestingRenderedComponent(
+          const menuItemComponent = mountWithProviderAndGetComponent(
             MenuItem,
             <MenuItem
               aria-disabled={disabledValue}
@@ -146,7 +146,7 @@ describe('MenuItem', () => {
 
     describe('as a Tab', () => {
       test('root role should be presentation', () => {
-        const menuItemComponent = getTestingRenderedComponent(
+        const menuItemComponent = mountWithProviderAndGetComponent(
           MenuItem,
           <MenuItem accessibility={tabBehavior} />,
         )
@@ -154,7 +154,7 @@ describe('MenuItem', () => {
       })
 
       test('anchor role should be tab', () => {
-        const menuItemComponent = getTestingRenderedComponent(
+        const menuItemComponent = mountWithProviderAndGetComponent(
           MenuItem,
           <MenuItem accessibility={tabBehavior} />,
         )
@@ -163,7 +163,7 @@ describe('MenuItem', () => {
 
       test('aria-label should be added to the child anchor', () => {
         const ariaLabel = 'Useful Tool Tip'
-        const menuItemComponent = getTestingRenderedComponent(
+        const menuItemComponent = mountWithProviderAndGetComponent(
           MenuItem,
           <MenuItem aria-label={ariaLabel} accessibility={tabBehavior} />,
         )
@@ -174,7 +174,7 @@ describe('MenuItem', () => {
 
       test('aria-labelledby should be added to the child anchor', () => {
         const ariaLabelledByID = 'element-that-labels'
-        const menuItemComponent = getTestingRenderedComponent(
+        const menuItemComponent = mountWithProviderAndGetComponent(
           MenuItem,
           <MenuItem aria-labelledby={ariaLabelledByID} accessibility={tabBehavior} />,
         )
@@ -188,7 +188,7 @@ describe('MenuItem', () => {
       const activeFlags = [true, false]
       activeFlags.forEach(activeValue => {
         test(`aria-selected should be ${activeValue} if menuitem active prop is ${activeValue}`, () => {
-          const menuItemComponent = getTestingRenderedComponent(
+          const menuItemComponent = mountWithProviderAndGetComponent(
             MenuItem,
             <MenuItem active={activeValue} accessibility={tabBehavior} />,
           )
@@ -202,7 +202,7 @@ describe('MenuItem', () => {
 
       activeFlags.forEach(activeValue => {
         test(`aria-selected should be ${activeValue} if menuitem also has the active prop and aria-selected prop is ${activeValue}`, () => {
-          const menuItemComponent = getTestingRenderedComponent(
+          const menuItemComponent = mountWithProviderAndGetComponent(
             MenuItem,
             <MenuItem
               aria-selected={activeValue}
@@ -220,7 +220,7 @@ describe('MenuItem', () => {
 
       test('aria-controls should be added to the child anchor', () => {
         const ariaControlsPanelID = 'panel-that-is-controlled'
-        const menuItemComponent = getTestingRenderedComponent(
+        const menuItemComponent = mountWithProviderAndGetComponent(
           MenuItem,
           <MenuItem aria-controls={ariaControlsPanelID} accessibility={tabBehavior} />,
         )